refactor: rename ingameCount to statusIndicator
